requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0
langgraph>=1.0.0
chromadb>=0.5.0
streamlit>=1.40.0
python-docx>=1.1.0
pypdf>=4.0.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
duckduckgo-search>=6.0.0
